requirements.txt
python-dotenv~=1.1.1
google-generativeai
protobuf~=5.29.5
google-auth-oauthlib~=1.2.2
google-api-python-client~=2.177.0